eshop/src/main/java/lt/codeacademy/eshop/service/MessageService.java
package lt.codeacademy.eshop.service;

import lombok.extern.slf4j.Slf4j;
import org.springframework.context.MessageSource;
import org.springframework.context.i18n.LocaleContextHolder;
import org.springframework.stereotype.Service;

/**
 * @author Andrius Baltrunas
 */
@Service
@Slf4j
public class MessageService {

    private final MessageSource messageSource;

    public MessageService(MessageSource messageSource) {
        this.messageSource = messageSource;
    }

    public String getMessage(String key) {
        try {
            return messageSource.getMessage(key, null, LocaleContextHolder.getLocale());
        } catch (Exception e) {
            log.error(String.format("Message not exist %s ", key), e);
        }

        return "";
    }
}
